+++ spec/lib/where_is_my_friends/local_topics_spec.rb
@@ -1,6 +1,4 @@
 # frozen_string_literal: true
-# rubocop:disable Discourse/Plugins/NamespaceMethods -- RSpec example-group helper.
-
 RSpec.describe WhereIsMyFriends::LocalTopics do
   fab!(:user)
   fab!(:category) { Fabricate(:category, minimum_required_tags: 2) }