+++ spec/models/where_is_my_friends_ai_provider_profile_spec.rb
@@ -1,6 +1,4 @@
 # frozen_string_literal: true
-# rubocop:disable Discourse/Plugins/NamespaceMethods -- RSpec example-group helper.
-
 RSpec.describe WhereIsMyFriendsAiProviderProfile do
   fab!(:admin)