+++ spec/requests/where_is_my_friends/interest_serialization_spec.rb
@@ -1,6 +1,4 @@
 # frozen_string_literal: true
-# rubocop:disable RSpec/DescribeClass -- Request-level serialization contract.
-
 RSpec.describe "Where Is My Friends interest serialization" do
   fab!(:viewer, :user)
   fab!(:member, :user)