+++ spec/requests/where_is_my_friends/practice_invitations_controller_spec.rb
@@ -1,6 +1,4 @@
 # frozen_string_literal: true
-# rubocop:disable Discourse/Plugins/NamespaceMethods -- RSpec example-group helpers.
-
 RSpec.describe WhereIsMyFriends::PracticeInvitationsController do
   fab!(:sender, :user)
   fab!(:recipient, :user)